Поправила часть замечаний после code review и добавила  api тесты

+++ migrations/000001_create_fabric_table.up.sql
@@ -2,8 +2,8 @@ CREATE TABLE IF NOT EXISTS `fabric`
 (
     id BINARY(16) NOT NULL,
     name VARCHAR (255) NOT NULL,
-    cost FLOAT NOT NULL,
-    amount FLOAT NOT NULL,
+    cost DECIMAL(10, 2) NOT NULL,
+    amount DECIMAL(10, 2) NOT NULL,
     created_at DATETIME NOT NULL,
     updated_at DATETIME,
     PRIMARY KEY (id)
